requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==42.0.5
    # via pyjwt
fastapi==0.110.1
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via anyio
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5
    # via jinja2
oauthlib==3.2.2
pycparser==2.22
    # via cffi
pydantic==2.7.0
    # via fastapi
pydantic-core==2.18.1
    # via pydantic
pyjwt==2.8.0
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.29.0
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
